Add OKX perpetual-swap live broker (ccxt)
New broker plugin for OKX USDT-margined perpetual swaps via ccxt.pro:
- src/broker/okx-broker.js — OKxDataSource + OkxExecutor + createOkxBroker;
  one-way (net) position mode, software SL/TP, cross/isolated margin, stall
  watchdog, --close-on-exit, --dry-run, --key-prefix for sub-accounts
- src/broker/okx/client.js — ccxt.pro wrapper: candle streaming with
  timestamp-transition closed-candle detection (OKX cache length stays 1),
  paginated history fetch, market order with retry, position/balance fetch,
  ensureFuturesAccountMode (skipped in --demo; acctLv checked at init in prod)
- src/broker/okx/intervals.js — OKX timeframe map
- src/broker/okx/okx-broker.test.js — 14 unit tests (all green)
- src/fetch-okx.js — history downloader to parquet (mirrors fetch.js for OKX)

Shared modules lifted out of tinkoff/ into broker/common/ so both brokers use
the same StateManager and PositionStore; tinkoff-broker.js import paths updated.

candle-parquet.js: volume column INT64 → DOUBLE (crypto volume is fractional;
lossless for MOEX integer volumes within double's 2^53 integer range).

package.json: ccxt ^4.5.58 dep; diviner-fetch-okx bin entry.

ml/train.py: update A05 description (sole ML filter, no gap pre-filter).

Live demo verified end-to-end on OKX sandbox 2026-06-16:
stream → signal → buy 0.01 BTC/USDT:USDT @ 65581.5 → SIGINT →
--close-on-exit reduce-only sell → all executed cleanly.

+++ src/core/candle-parquet.js
@@ -11,7 +11,12 @@ export const candleSchema = new parquet.ParquetSchema({
     high: { type: "DOUBLE" },
     low: { type: "DOUBLE" },
     close: { type: "DOUBLE" },
-    volume: { type: "INT64" },
+    // DOUBLE (not INT64) so fractional volumes are representable — crypto venues
+    // report sub-unit base-currency volume (e.g. 14.3326 BTC). MOEX share volumes
+    // are whole numbers and stored exactly within a double's 2^53 integer range,
+    // so this is lossless for the Tinkoff path too. Readers normalise via
+    // Number(), and existing files keep their own embedded schema on read.
+    volume: { type: "DOUBLE" },
     // Interval length in minutes (1, 60, 1440, …). Optional so a single-interval
     // file can omit it (legacy/back-compat: readers treat such rows as the base
     // interval from metadata). A multi-interval file tags every row, letting the
